Organize component files into dedicated folders
