+++ src/components/Pricing/Pricing.css
