+++ src/components/Services/index.js
@@ -1,0 +1,1 @@
+export { default } from './Services.jsx'
